BugId: GS886 LeaseManager now schedules execution at a fixed rate.
In fixed-rate execution, each execution is scheduled relative to the scheduled execution time of the initial execution. If an execution is delayed for any reason (such as garbage collection or other background activity), two or more executions will occur in rapid succession to "catch up." In the long run, the frequency of execution will be exactly the reciprocal of the specified period (assuming the system clock underlying Object.wait(long) is accurate).

svn path=/trunk/gigaspaces/; revision=6242


Former-commit-id: bcda99df5cbc03760080392b7ac16fa6356951c6