BugId: GS948 The /./nameOfSpace_container/nameOfSpace is not a legal space url. It needs to have a ?schema= attribute since If using the /./ notation it appends the ?schema=default.
And when using the /./ you should NOT add the container name i.e. nameOfSpace_container.

In that case the bug was in the Space URL parser which was left with the following space name
nameOfSpace_container/nameOfSpace
I have added the support for it and therefore you want be getting to the point where the log throws exception since there is no such directory nameOfSpace_container/nameOfSpace....

svn path=/trunk/gigaspaces/; revision=6262


Former-commit-id: d8a34f346cb0cda3b2497d5b974019d5cb570ace