BugId: GS1066 Enable download of properties file via HTTPD/web-server, e.g. for ?properties spaceurl attribute. Now you can pass one of the following SpaceURL's in order to load the gs.properties file (used for overwrite config settings):
Either (searched in a location under web server's url):
&properties=http://<host>:<port/lib/gs.properties --> you can put the file under the Webster http root dirs...

Or as before (searched under <some root>/config folder and if the <some root> is in the -classpath):
&properties=gs

Now you can put the properties file under an accessible single http url and all spaces will download it from this path. The local search via the classpath definition is also supported as before.

svn path=/trunk/gigaspaces/; revision=6491


Former-commit-id: 1ca9b006115a2d1eb18779246ad4ceb5159b77c2