BugId: GS1115 The <min_ext_index_activation_size> is missing. This parameter is the threshold where index is maintained ? i.e. whenever there are more than <min_ext_index_activation_size> amount of entries stored within the space an extended index will be maintained.
We might have another schema that called jdbc, or query to include <enabled-classes>*</enabled-classes>
Here is the full description of the <extended-match>
An entry's attributes can be indexed using explicit or implicit indexing (see Indexing). By default, when class attributes are indexed, a hash-based index is maintained. The hash-based index is used for the regular JavaSpaces matching (i.e., "equals") and accelerates the read () and take () operations. In order to accelerate the read () and take () operations that use SQLQuery with greater than/less than operations, turn on the btree index.
To turn on the btree index, modify the space configuration file as follows (note the asterisk):
<extended-match>
         <enabled-classes>*</enabled-classes>
         <regular-expressions-cache-size>300</regular-expressions-cache-size>
</extended-match>
By default the btree index is not maintained for indexed class attributes. Leaving the btree index turned on impacts the write performance.
The <enabled-classes>*</enabled-classes> can be replaced with a list of class names separated with commas that you want to index using the btree index
For example: <enabled-classes>classA,classB</enabled-classes>

Now the cache, persistent and default space schemas contains the following block:
<!-- An entry's attribute can be indexed using explicit or implicit indexing.
           		By default, when class attributes are indexed, a hash-based index is maintained.
           		The hash-based index is used for the regular JavaSpaces matching (i.e., "equals")
           		and accelerates the read () and take () operations. In order to accelerate the read () and take ()
           		operations that use SQLQuery with greater than/less than operations, turn on the btree index.
           		By default the btree index is NOT maintained for indexed class attributes (no value in <enabled-classes> )
           		Leaving the btree index turned on impacts the write performance.
      				To turn on the btree index, modify the space configuration file as follows (note the asterisk):
      				<enabled-classes>*</enabled-classes> or <enabled-classes>classA,classB</enabled-classes>
           		-->
                <extended-match>
                     <enabled-classes></enabled-classes>
                     <!-- This parameter is the threshold where index is maintained  i.e. whenever
                     there are more than <min_ext_index_activation_size> amount of entries stored within the space
                     an extended index will be maintained.
                     Default value: 10 -->
                     <min_ext_index_activation_size>10</min_ext_index_activation_size>
                     <!-- Default value: 300 -->
                     <regular-expressions-cache-size>300</regular-expressions-cache-size>
                </extended-match>

svn path=/trunk/gigaspaces/; revision=6679


Former-commit-id: 00298daeeca75a3765f24488d3319678b3234430