BugId: GS974 * removed partitioned-master-local cluster schema. * modified partitioned-cluster and partitioned-sync2backup with notify policy-type to support broadcast notify.
To enable broadcast notify, we set the load-bal-policy to its default or explicitly set:
<notify>
	<policy-type>hash-based</policy-type>
	<broadcast-condition>broadcast-if-null-values</broadcast-condition>
</notify>

To disable explicitly set <broadcast-condition> to broadcast-disabled.

In a replicated master-backup topology, <replicate-notify-templates> should be set to true,
and <trigger-notify-templates> should be set to false.

A notification will be registered in each of the (partitioned) groups seamlessly.
If failover occurs, the backups will then trigger the notifications.

svn path=/trunk/gigaspaces/; revision=6990


Former-commit-id: 682b1e92d596ab9324afd514b69ae00178ace000