BugId: GS1107 useLocalCache/-cache parameter as part of benchmark arguments has been changed to adapt to latest product equivalence.
When passing useLocalCache/useLocalCache=true as part of the space url argument, -cache is turned on.
When passing useLocalCache=false as part of the space url argument, -cache is set to false.
When -cache parameter is passed, and no contradicting space url argument, then -cache is turned on.

This will conform to:
-cache ==> DCacheSpaceImpl via SpaceFinder
-map ==> GSMapImpl via CacheFinder
-map -cache ==> MapCache via CacheFinder

*SpaceUrl changed to allow null arrtibute value in URL, when value is not manditory.

TODO:
Browser benchmark Master-Local radio buttons should be removed.
Only have use local-cache checkbox.

svn path=/trunk/gigaspaces/; revision=7193


Former-commit-id: ba2c0ab0426d5b7c59d92a0fe2acef8a0b030d0a