BugId: GS1255 Validate Space url string for a valid setting of attributes, their values, their type (boolean, int, String etc.). On Each call to SpaceURLParser a validation of the url string will be performed. If the validation fails a warning will be thrown. Next step will be to validate that for specific attibutes, such as cluster_schema, the related attributes must be set as well (i.e. id, total_members), otherwise these should fail in validation as well.
We use xsd schema to validate the space url attributes. The spaceUrlXML-Schema.xsd is located in the JSpaces.jar. The internal space url attribute ignoreValidation is used to bypass the validation process (e.g. for the AbstractReplicator and other methods which internally look for members)

Add ignoreValidation attribute in clusterAxml file

svn path=/trunk/gigaspaces/; revision=7258


Former-commit-id: 34774aabd1ae925b112ad03cef1955930ecf8070