BugId: GS1288 The FIFO type indicator in the type-directory is always determined by the first operation to introduce it, regardless of the operation itself. Thus, a read/take operation can introduce a type with FIFO true, while a subsequent write is performed with FIFO false. In such a case, an InvalidFifoClassException will be thrown.
Resolution:
InvalidFifoClassException and InvalidFifoTemplateException should retain their original logic.
InvalidFifoClassException should not be thrown when a read/take with FIFO=false is performed
prior to a write/snapshot/update with FIFO=true. i.e. the FIFO class-type indicator will be determined
only when introducing a class by a non-template operation.

svn path=/trunk/gigaspaces/; revision=7428


Former-commit-id: b21b1c6c87374a22b4a31302e2e6753f831c313b