BugId: GS1481 The gigaspaces.config (jini service config file) contains hard coded values. The space, schema, spaceurl etc. were set hard coded inside the /services/gigaspaces.config and since this file is used by the gsServer shell, it ignored the SpaceURL passed/set in gsServer.
Now the config file contains the config elements marked and the spaceURL contains the variable placement ${gs.space.url} which is set also by the gsServer.

Raised by a support call of CME who tried to pass gsSever "/./mySpace" but it insisted on using the myCache values from the gigaspaces.config.

svn path=/trunk/gigaspaces/; revision=8267


Former-commit-id: 45582e2761c0ac9b789d20b9ea38ce4a37640a13