BugId: GS1489 Added the support for the SpaceURL "create=false" attribute. Until now we checked only if the "create" was passed and did not support "create=false".
Now, if it is passed we check in the containers map according to its container name and if such container already exists we return the collocated space proxy if such already exists.

Note, every time we create new space we check if such container already exists in the containers weak map, according to its SpaceURL object equals. In the "create=false" case we do not care what is the passed full SpaceURL so we check only for the name. Meaning the next time you will ask for the same collocated space, we do not care about the schema or other SpaceURL attributes.

svn path=/trunk/gigaspaces/; revision=8322


Former-commit-id: 7adf22e085a861561afbcc2a487a22f23af53350