BugId: GS1507 The helloSpacePojo examples shows the basic JavaSpaces API - read, write, take, notify, notifydelegator and update - as well as transactions and iterators.
The only difference of this example comparing to the helloSpace, is the usage
of Pojos (which being converted to Jini Entry) instead of an Entry.

The example converts pojos to entries and vice versa using a pojo2entry converter utility.
Please note the pojo can have a meta data mapping using a declarative
approach through *gs.xml mapping file, or using declarative approach through Java Annotations.

In later releases we will have support for Pojos in the IJSpace
interface itself, avoiding using the utility.

svn path=/trunk/gigaspaces/; revision=8410


Former-commit-id: fd1a4d1033740e42db8ab83eee11293e185da17a