APP-12 Support compression og JMS Text message body. Using configuratin one can set teh minimum message size from which we will compress teh message body.
Used CompressedMarshObjectConvertor to compress the Text message body when sent and decompress it when message is received.
Added to GSJMSAdmin a field which is set by the user, and defines the min message size in bytes from which we make a compression.
e.g. if a 1 MB Text JMSMessage body is sent, and the compressionMinSize value is 500000 (0.5MB) then we will compress that message body (only), otherwise we will send (write) it as is.

TODO Currently this configured via system property -Dcom.gs.jms.compressionMinSize=500000
later on we will add it to the JMS-config.xml
The default min value is 0.5 MB

svn path=/trunk/gigaspaces/; revision=8603


Former-commit-id: 45581f591f8906c41118a3ee526aa7075c95a29c