APP-109 Modify the gsInstance to use a -classpath by default and not the -jar.
Since it caused many confusions in the past few months both from internal users, dev , support and customers.
When using -jar you cannot modify the classpath nor add the /config first in classpath, that is very often then the users expect that the schemas will be loaded from the file system but instead they are loaded from the jar.

More sadly Is even both options exits and well documented in the batch file still the users are not opening it to review.

svn path=/trunk/gigaspaces/; revision=9176


Former-commit-id: 0da3491667ee645f75026859c344a245508b54e3