CORE-409
Primary key is incorrect when POJO's gs.xml is defined with multiple indexes.

Bug:
The TypeTableEntry is built with first-defined-index regradless of the declared primary key. This is the correct approach
only if no primary key was declared. Otherwise, the primary key should hold. This affects the EntryPacket that is built
from the TypeTableEntry.

Fix:
The TypeTableEntry should extract the primaryKeyName from the ExternalEntry as assigned by the Converter.
Otherwise, rely on the first defined index (or first field if no indexes are declared).

Reviewer: Guy K.

svn path=/trunk/gigaspaces/; revision=10120


Former-commit-id: 76f5fdd90825ae4eebe44704db9ccb61d09cef3a