APP-495 problem with generateGsXml: missing echo in script caused:
When running generateGsXml.bat I'm getting:
'a.' is not recognized as an internal or external command,
operable program or batch file.
'b.' is not recognized as an internal or external command,
operable program or batch file.
'take' is not recognized as an internal or external command,
operable program or batch file.

svn path=/trunk/gigaspaces/; revision=10379


Former-commit-id: e00da0135b1238e75f4d49a61cbb86eb5047d63a