CORE-516
Bug:
When using NotifyDelegatorMultiplexer on a partitioned cluster, BroadcastNotify registration returns
a fictive EventRegistration with a fixed eventId. This prevents from notifications to arrive at the client
listener.

Fix:
Return the last event id registered, which will serve as the id for the entire partition.
Although this doesn't fix the problem entirely, it does enable notifications to be received at the
client listener.

This issue still needs proper investigation.

svn path=/trunk/gigaspaces/; revision=10990


Former-commit-id: 5be1e89e8227aebe27a357605d82f8b0507fa1ff