APP-732 When providing the UID, the lookup was done based on the value and not the fact that it is a UID. This resulted in the fact that we do not get EntryNotInSpaceException and a mismatch between Entry and Pojo behaviour.
The reason we did this was because we did not have the SpaceRouting index option, and users wanted to lookup based on the primary key that acted as the partition hash code. Now that we have SpaceRouting, the user has the option to define a UID and on a differnt property define the SpaceRouting.
rev:shay b

svn path=/trunk/gigaspaces/; revision=11420


Former-commit-id: 369ec9465e44bb405defcb3291210f7212afea1c