CORE-568
NullPointerException is thrown when using readEnvelope and a cached-persistent entry.

Bug:
using readEnvelope, if the entry resides in cache, the SA context is never acquired, causing a NPE.
Only when the entry is not in cache, and we are using an LRU persistent space, the context is acquired.

Fix:
acquire the context only when using LRU persistent space and disregard its existence in cache.

Reviewer: Y.

svn path=/trunk/gigaspaces/; revision=11570


Former-commit-id: 48f4bf5bf9f200e519f48eefe4a5005e76ba75a8