APP-220 Remove the if "%TITLE%" == "" (
which caused to a break of the url string and messages such as this one:

Starting a Space Instance
Setting space url to "/./myCache?schema=default&properties=gs&groups="gigaspaces-5.2EE""
'properties' is not recognized as an internal or external command,
operable program or batch file.
'groups' is not recognized as an internal or external command,
operable program or batch file.
Reviewer: Shay H.

svn path=/trunk/gigaspaces/; revision=11666


Former-commit-id: 24f1297874682aa473b223ee8de1cb8a80836d63