PACK-104
com.j_spaces.examples.jdbccachestore CacheStoreImpl (for both map and space) sometimes calls getConnection()
but doesn't always close it afterwards. This causes too many connections to be opened. This is noticeable on linux
machines since there is a defined limit to the number of connections that can be opened simultaneously.

svn path=/trunk/gigaspaces/; revision=11903


Former-commit-id: b32c3e730ac6bb7dc7e32e593997223b5803fc60