GS-1472 GS-1472 GS-1470 GS-1469 GS-1468 GS-1237 GS-1470 GS-322 1. Support running an embedded Mahalo Jini Transaction Manager when container starts, same as with Reggie. The need here is to have the Mahalo, Reggie, Webster and space running collocated (later using same GenericExporter) and set with same codebase (same httpd port ) etc. Also, since the gsServer was deprecated, we need a way to run gsInstance with Mahalo, Reggie, Webster and space running collocated. The Mahalo should NOT be loaded by default but configured using the container schema and system property overwrites.
Configuration changes:
- The container schema files has a new section for enabling the embedded Mahalo tx manager inside the space:

<embedded-services>

...

<mahalo>
    <!-- If true, will start an embedded Mahalo Jini Transaction Manager. Default value: false -->
    <start-embedded-mahalo>${com.gs.start-embedded-mahalo}</start-embedded-mahalo>
</mahalo>

2. Support loading configuration resources without using classpath but scanning the file system under a home base.
The resource loader will try to load the resource using the classpath and if not found, for example when running in the IDE, As a fall back, the system will use the Locator.derivePath() which will scan recursively all directories under the home base to load the file.

3. Rename the gigaspaces.policy that resides inside the JSpaces.jar to be security.policy as in the zip.
The resource loader will try to load it from the zip root/policy/policy.all and if not found, for example when running in the IDE, it will fetch it from the JSpaces.jar from same path.
As a fall back, if the resource loader failed to load the security policy using the classpath the system will use the Locator.derivePath() which will scan recursively all directories under the home base to load the file.

4. Support overriding the webster current anonymous port when space or browser are started.
Currently the Webster starts embedded in the space, always with a generated anonymous port which the user has no control on and cant get its value.
There are scenarios (e.g. when the codebase is set explicitly and you need to define the port) that the user should be able to explicitly set the requested httpd port to be used all across the system.
Add support in container schema config xml and a system property overwrite.

Configuration changes:
- The container schema files has a new section for setting the webster port:

<embedded-services>
<!-- Starts a Webster HTTP Server embedded in the container VM
Default value: true -->
<httpd>
...

<!-- The host port the server socket HTTPD (Webster) is using is bound to.
Default value: port 9813 but if that is busy it generates an anonymous port-->
<explicit-binding-port>${com.gs.embedded-services.httpd.port}</explicit-binding-port>
<httpd>

5. Support overriding the java.rmi.server.codebase when space or browser are started. If the -Djava.rmi.server.codebase was set externally than we use it as is and we do NOT append to it the gigaspaces platform jars.
Before, we appended the JSpaces-dl.jar/JSpaces.jar, reggie-dl.jar (if reggie is used) and jsk-dl.jar jars to the codebase. Specifically the issue raised with the JSpaces.jar being downloaded using codebase, even if we did not want it to happen.
If the codebase was not set externally (the default behavior) we set it and add the JSpaces-dl.jar/JSpaces.jar, reggie-dl.jar (if reggie is used) and jsk-dl.jar jars to the codebase. If the codebase is set in the browser, then we also append the JSpaces-ui.jar.

6. Remove the need for policy file
7. Rename the gigaspaces.policy that resides inside the JSpaces.jar to be security.policy and support loading without using classpath
8. Support shallow product jars structure
Reviewer: Evgeny

svn path=/trunk/gigaspaces/; revision=13039


Former-commit-id: 5d7f9d4f580526badb3d01f52ccbb1af0e70d0d6