GS-1511
Manage redo-log for async-targets of the sync-backup space in order to retransmit them to async targets - forming a reliable path if primary fails.

Duplicate packets are transmitted (after recovery process) to complete the redo-log of the backup with
unconfirmed packets belonging to async-targets.

While the primary is active, it will forward ACKs from async-targets to this backup removing these packets from the redo-log.

svn path=/trunk/gigaspaces/; revision=13417


Former-commit-id: 2d91e008aeb3b68a94f114c083c86c00ee8577a3