GS-1582 Force the rmi to search for JSpaces jars first in classpath before downloading from rmi codebase Make the java.rmi.server.RMIClassLoaderSpi use the default implementation and not the Jini preferable CL.
As you know, the JSpaces-dl.jar (in 5.2) was downloaded with all its dependencies from EVERY space in the cluster causing dozens of MBs of jars going through the wires and overloading the HTTPD, Reggie and the space...

So, what we will do in 6.0 (and in next 5.2.x patch) is also change the Jini behavior by modifying the jsk-resources.jar which comes with the product and it does the same trick as the sys prop.

set RMI_OPTIONS=-Dsun.rmi.dgc.client.gcInterval=600000 -Dsun.rmi.dgc.server.gcInterval=600000 -Djava.rmi.server.RMIClassLoaderSpi=default -Djava.rmi.server.logCalls=false

To test it: just use the new jsk-resources.jar and replace the one under lib/jini.
- Turn on Webster logging that will show you the jars downloading in the Space side - org.jini.rio.tools.webster.level = ALL
See that the JSpaces.jar is no longer downloaded via http, while before this change it did twice for each space.
Reviewer: Igor

svn path=/trunk/gigaspaces/; revision=13428


Former-commit-id: 1acf61447db3bd43ed388c6199ae73ca7068ca55