GS-1719 ResourceLoader should always attempt to search ALSO using the resource file name only, in all class loaders and Locators.
Right now the ResourceLoader works in the following order:
If the passed resource string is config/schemas/yyy.xml
1. It calls the (current thread) contextClassLoader.getResource(config/schemas/yyy.xml).
if not found -->
2. It calls this classLoader.getResource(config/schemas/yyy.xml) (e.g. if running in AppServer)
if not found -->
3. Uses the com.gigaspaces.start.Locator.derivePath(yyy.xml) that searches recursively under the GS, using root file system.

The above 3 steps are to cover all scenarios, including different class loader and a flat structure resources (all in same directory).

BUT, we need to search in each step also using the resource name itself e.g. yyy.xml if the search using config/schemas/yyy.xml failed, since that is the most common scenario because the out of the box classpath structure is using config/schemas/yyy.xml.
- Without doing that we might get to step 3 in many times, which is the fall back and most expensive scenario.

The steps should be:

1. It calls the (current thread) contextClassLoader.getResource(config/schemas/yyy.xml).
if not found -->
1.1 It calls the (current thread) contextClassLoader.getResource(yyy.xml).
if not found -->
2. It calls this classLoader.getResource(config/schemas/yyy.xml) (e.g. if running in AppServer)
if not found -->
2.1 It calls this classLoader.getResource(yyy.xml) (e.g. if running in AppServer)
if not found -->
3. Uses the com.gigaspaces.start.Locator.derivePath(yyy.xml) that searches recursively under the GS, using root file system.

Reviewer: Evegny

svn path=/trunk/gigaspaces/; revision=13929


Former-commit-id: a4115de93cf1dc761ee847d9c6961c3a38e57303