GS-1679 FinderException is thrown when lookup is performed prior to the space being published but before timeout expires. Even when the LookupFinder was blocked (with the default timeout forever) waiting for notifications about newly added/registered/joined services, it never returned with an answer. That was caused by the FinalizerThread in LookupFinder (which checks if the LookupFinder is not active for a minute) which cleans the associated SDM, LDM and local cache and calls their terminate().
The LookupFinder creates and maintains a LookupCache as well as the SDM (that contains its own local cache). The problem happens when the FinalizerThread in LookupFinder calls the SDM and LDM terminate() since that causes to the cleanup also of "our" local cache with its listeners from LDM --> causing that when a new service is added no events will be sent to serviceAdded().
The fix was that the FinalizerThread will clean only the local cache and terminate it but will not terminate the local SDM and LDM.

That fix should enable the work on the UnitTestLauncher in a forked VM (remote) mode.
Also, it should fix the issues with the new TG clustered tests that are based on the UTL.
Reviewer: Misha

svn path=/trunk/gigaspaces/; revision=14161


Former-commit-id: 8b0a5235c7a1d5800ead67dda9d6663b9daa83d7