GS-1828 Reduce number of calls to deepClone() while creating cluster policy.
The ClusterXML.java is calling 3 times (for LoadBalancingPolicy, for ReplicationPolicy and for FailOverPolicy objects) to IOUtils.deepClone(), for EACH cluster policy that is created PER EACH space member, which is VERY expensive.
If you have a 50 nodes cluster it is called 2500 times * 2 which takes 40% of the time to create the all cluster config in memory

Aditional change was, not to call deepClone() at all while setuping the cluster config but only on demand. Also, before we used to call it in both places, when first cluster config created and then later when FO policy tried to lookup the members.
Reviewer: Evgeny

svn path=/trunk/gigaspaces/; revision=14205


Former-commit-id: 914e6692b256204333c00518bed6a399ba8c9b04