GS-1851
Use Case:
Entry which resides in external-data-source isn't async replicated.

Resolution:
Keep in memory redo-log the metadata info needed to fetch when replicating the entry. This is relevant when the entry no longer resides in cache, and its information is fetched from the underlying data-source.

Till now, this was done only for TAKE operations. This is now extended for all operations when using LRU.

When replicate-original-state or sync-replication is enabled, we still keep the full packet value in redo-log regardless of the above.

Reviewer: Guy K.

svn path=/trunk/gigaspaces/; revision=14530


Former-commit-id: dfe011b066b7ff9a086582b95b3457794b4584a9