GS-2422
Summary:
Primary doesn't replicate delta of other async targets into backup when there is nothing new to replicate to the backup.

Resolution:
When beginSession is called from recovering backup, it forces a connection check at the primary to reconnect if ping failed.

Note: This does not take into account a disabled transsmission matrix. If a backup does not replicate into the primary, it will never invoke beginSession, and the primary will keep the old proxy until it will have something to replicate.

Reviewer: Igor g.

svn path=/trunk/gigaspaces/; revision=15788


Former-commit-id: 14f84439579c30b7e0e9899b59f84ca713c8d680