GS-2494
Start additional LookupService cause to SpaceBrowser to remove and then add space.
JoinManager of the space received as argument not immutable proxy it means that two or more different modules like WorkerManager FilterManager changed the proxy state during the join-manger did his job. So when new Lookup got event from JoinManager that the registered proxy is not from the same version, because LUS check the version with byte-comparing.
Solution is to pass to join-manager the proxy that no one can't change the state.

svn path=/trunk/gigaspaces/; revision=16049


Former-commit-id: 1edcc4d1e3b40f6fe522fdbf514602b011c90d35