GS-2216 In a non-multicast environment, it appears that a client may not "re-acquire" a lus after all lus's have been killed.
The problem:
---------------
In a non-multicast environment, it appears that a client may not "re-acquire" a lus after all lus's have been killed. This makes it impossible to find the new services that register with the new lus. There were 2 underlying causes:
- It appears that a reference to a lus that has become unreachable is not discarded quickly. The lus reference must be discarded from the internal cache before re-discovery will even be attempted.
- Also, re-discovery attempts were done less frequently as time passed. Initial attempts were made at 5s, 10s, 20s, etc.. but before long discovery was only attempted every 15 minutes. This would be an issue if the time between the crash of the lus and the recovery was several minutes because after several minutes, the frequency of discovery attempts is unacceptably low.

The fix in short:
-------------------
The correct location for the patch would be net/jini/discovery/LookupLocatorDiscovery, which is responsible for executing unicast lookup discovery and also for keeping each lus "discovered".
- Modify the class to actively "ping" the lus at a configurable interval. If the lus is unreachable, it is "discarded" which is a prerequisite to re-discovery.
- The lus discovery "schedule" was hard coded and would become too infrequent after a few minutes.
New setting was added for the discovery schedule by setting using a "-D" option.

Detailed explanation:
-------------------------
The LookupLocatorDiscovery modifications provides the ability to periodically
test the reachability of discovered lookup service instances, as well as
providing the option to declare retry intervals used to re-establish
communication with a failed lookup service.

This approach included in this project provides a more responsive unicast
discovery approach by declaring the interval(s) the utility will sleep between
unicast discovery attempts (or retries), and also provides a means to verify
lookup service availability by periodically obtaining information from the
backend implementation.

The default behavior of the LookupLocatorDiscovery utility has been modified to
create an internal DiscoveryListener that will create a LookAliveTask per
discovered LookupService instance. Each LookupAliveTask will periodically invoke
the ServiceRegistrar.getLocator() method (resulting in a round trip to the
backend implementation). If a non-recoverable exception occurs (ConnectException
or otherwise) the LookAliveTask will discard the ServiceRegistrar in question.
The periodicity of the LookupAliveTask can be controlled by the
com.gigaspaces.unicast.interval system property. The default interval is
1 minute.

Example :

-Dcom.gigaspaces.unicast.ping=5000

Will cause the LookupLocatorDiscovery utility to wait 5 seconds between
invocations of the getLocator() method.

The LookupLocatorDiscovery unicast retry (sleep) intervals have been untouched.
If the user wishes to tune the unicast retry intervals, the
com.gigaspaces.unicast.interval system property is used to control the
behavior of this LookupLocatorDiscovery utility. A comma separated list of values
defining the intervals to wait between subsequent retries. Values are declared
in milliseconds.

Example :

-Dcom.gigaspaces.unicast.interval=5000

Will cause the LookupLocatorDiscovery utility to wait 5 seconds between retries

-Dcom.gigaspaces.unicast.interval=5000,10000

Will cause the LookupLocatorDiscovery utility to first wait 5 seconds, then
10 seconds between retries. This declaration provides a graduating approach
(similar in approach to the default settings), increasing the amount of time to
wait before the next discovery attempt should be made - upon each invocation,
eventually reaching a maximum time interval over which discovery is re-tried.
In this way, the network is not flooded with unicast discovery requests
referencing a lookup service that may not be available for quite some time
(if ever).

- Tested using attached use case

svn path=/trunk/gigaspaces/; revision=16475


Former-commit-id: 9d74d6fe341ff2425c3a27bea9443e62b77c42cc