GS-2733
for dirty-read or read-committed or read not under txn, same entry can return more than once due to scanning of shadow entries, so we verify uid uniquness. for LRU, eviction is handled by EntriesIter.

Fix:
in ALL_IN_CACHE, also keep a returned list of uids for entries read which are not under txn.
in LRU, EntriesIter should not return the entry if already returned.

Reviewer: Yehiel.

svn path=/trunk/gigaspaces/; revision=17070


Former-commit-id: 7d17f922ab9fcedab4fec5599438c222916f8e20