GS-2692
iterator.cancel() may be called directly or indirectly by iterator lease expiration. When called by the latter, we only check if it was terminated - which indicates expiration or cancellation. We should verify cancellation since it may be the case where lease termination does not cancel the iteration.

To properly handle the case of blocking next (see GS-2692) we will notify the client even if the iteration was already cancelled by the end of the lease timeout.

svn path=/trunk/gigaspaces/; revision=17206


Former-commit-id: fefded338aaeaf247d89d4be7b52ec01324fbc82