GS-2781
The problem was in the way JMS messages instanciate. I change their constructors so the objects could be copied properly through the network.
I also removed the code that sets the EntryInfo uid to the message ID. From now the UID is set automatically by the space.
The JMS passed successfully a full examination in CTS.

svn path=/trunk/gigaspaces/; revision=17213


Former-commit-id: e8da048f22cb31967cec8f0d3b36f14b2abcc18c