GS-482 Perform xsd validation on the correctness of values entered in the schemas. The validation of xsd is relevant only when xsl is used (e.g. when using semi-dynamic clustering), since when a static cluster xml is used the xsd validation is used.
Note - The previous clusterXML-Schema.xsd schema that is loaded from the JSpaces.jar and used only for the static cluster xml file, is going to be used now also for the validation of the semi-dynamic (xsl) cluster creation.

In addition the clusterXML-Schema.xsd as is structured today is not updated with all elements and needs also to guide the end user with the valid values and types each element can have.

Note: the schema validation enablement is turned on by default so there might be cases when users built their own customized xsl files but they failed to pass validation. In such case, the users will be instructed to fix the xsl files since in any case what they built was invalid in first place and would have fail in runtime.

There was always a flag to turn off the cluster schema validation setting:
-Dcom.gs.xmlschema.validation=false
Reviewer: Kobi

svn path=/trunk/gigaspaces/; revision=20346


Former-commit-id: 871499324471fed4ca793a53f70d43677401a8bc