GS-3953: Direct proxy extracted from clustered proxy, holds space config filled with ClusterPolicy.
The JSpaceContainerImpl stores a container table with all proxies to each cluster member.
These direct proxies contain a Space-config extracted from JSpaceImpl. The Space-Config is that of the cluster, and thus contains also the cluster-policy.

When extracting a direct proxy (embedded-proxy in cluster-proxy) it's space config should not contain 'cluster' meta-data. The new getEmbeddedProxyConfig() extends the functionality provided by getConfig() and removes unrelated meta-data, such as cluster-policy.

Reviewer: Shay B.

svn path=/trunk/gigaspaces/; revision=21946


Former-commit-id: 2fa3d8a2130fdbdb5c647ca4634480a5d9e95a73