GS-4165: Externalizable's version is always written to the stream.
write/writeMultiple are affected by the version field always being written to the stream.
If the version is zero, we don't need to write it. Now, getVersion() always tries to extract it from
the object first, and only when zero from the outer wrapper.

Due to current construction issues - we carry the version in the outer wrapper and
not inside the user object.

Reviewer: Guy K.

svn path=/trunk/gigaspaces/; revision=22986


Former-commit-id: 39a465064f21c1426d6015f03475215bc4af8d4f