GS-153 writeMultiple semantics Changed from last version:
In IResult type was renamed to ResultType, getResult() was renamed to getResultType() and getLease() now return Lease and not Object.



public interface IResult {



      public static enum ResultType

      {

            LEASE, ERROR

      }



      public ResultType getResultType();



      public Exception getException();



      public Lease getLease();



MultipleOperationPartialFailException was renamed to WriteMultipleOperationPartialFailException

With one method   public IResult [] getResults() (note, the return type is array instead of collection).



The new commit should fix as well one more path, the case where the query is scatter and gather between spaces.

Reviewer: Yechiel

svn path=/trunk/gigaspaces/; revision=24306


Former-commit-id: bb5d2f3d8e277282e87cbb5a2a8e9fd1a7ec4caa