GS-4504: gsc's remains with the same memory after relocation.
1.
When deploying a single space:
The FilterProvider holds a SpaceRemotingServiceExporter$RemotingServiceInvoker which holds a reference to a space -
The filter is part of the custom properties used to initialize the space. Since the JProperties is a static map, this reference is not removed.
- Fix: cleanup JSpaceAttributes, customProperties, dcache properties.

2.
When deploying a clustered space:
The FilterProvides (like in previous case) is held by the SpaceURL, which is part of the clustered proxy LB proxies.
- Fix: cleanup the cluster-proxy and embedded-proxy references.

Reviewer: Shay B.

svn path=/trunk/gigaspaces/; revision=25726


Former-commit-id: 4b7cbb4a438b94a1f921183d5b1a1634368882a8