GS-4733: service-failure event may cause duplicate re-deployments.
A GSM may receive serviceAdded events on services which can no longer be reached (usually after failure).
If this happens, it starts to manage a fault-detection handler against this service - only to find out (after 3 retries) that it does not exist.
This initiates a serviceFailure event, which may cause under certain circumstances to deploy an instance of this service.

Fix - when serviceAdded is called for an unreachable service, don't manage it.

svn path=/trunk/gigaspaces/; revision=26555


Former-commit-id: 2d0b053523dd14145053762bac0f0b7f315eb0d4