GS-4770: Relocation of a service may not obay max-per-vm.
When relocation of a service is performed, a sequence of destruction is performed against the service.
It may happen that in the process, the fault-detection handler of the service may also recognize that the service is no longer reachable,
initiating a sequence of service-failure. Carrying out these two sequences may lead to duplicate provisioning of a service on two different GSCs.

Fix:
When service-failure is carried out, it ignores the 'pending' state of services and starts dispatching provision requests.
Thus, we should only provision when actual+pending < maintains.

Reproducible in builds: 6.0.5-2149-2, 6.5-m11-b2330, and	6.5-2330-65

Reviewer: kimchy

svn path=/trunk/gigaspaces/; revision=26646


Former-commit-id: 7a5f5ec5e0f701d13a1c695f0c327bc2da0eedd0